==> notebooks/5.0-yp-train_model_SARSA.ipynb <==
import os
print(os.getcwd())
def update_working_directory():
    from pathlib import Path
    p = Path(os.getcwd()).parents[0]
    os.chdir(p)
    print(p)
update_working_directory()

import numpy as np
from tqdm import tqdm
import json
import sys

from src.data.env_dog import get_env_actions, get_env_space, env_reset, get_happiness, env_step

from src.features.rl_tools import init_Q, get_epsilon, get_alpha, epsilon_greedy
from src.features.rl_tools import init_steps_per_state, update_steps_per_state
from src.features.rl_tools_during_training import get_default_checking, launch_checking, save_models, display_analyse_Q

from src.models.value_based.sarsa.rl_tools_sarsa import update_Q_SARSA

from src.utils.plots import get_info_params
from src.visualization.rl_plots_evolution import plot_evolution_reward, plot_evolution_steps
from src.visualization.plot_happiness import plot_evolution_happiness

method_id = 'sarsa_1'

# Parametrisation
params = json.loads(open('src/models/value_based/sarsa/{}.json'.format(method_id)).read())

info_params_dict = {
    'method': params['method']
    , 'method_id': method_id
    , 'n_episodes': 0
    , 'gamma': params['gamma']
    , 'nmax_steps': params['nmax_steps']
    , 'init_Q_type': params['init_Q_type']
    , 'start_at_random': params['start_at_random']
    , 'init_epsilon': params['epsilon']['init_epsilon']
    , 'decay_epsilon': params['epsilon']['decay_epsilon']
    , 'min_epsilon': params['epsilon']['min_epsilon']
    , 'init_alpha': params['alpha']['init_alpha']
    , 'decay_alpha': params['alpha']['decay_alpha']
    , 'min_alpha': params['alpha']['min_alpha']
}
info_params = get_info_params(info_params_dict)


info_params

# Initializing environment
action_dict, n_actions = get_env_actions()
_, n_states = get_env_space()

# Initializing the Q-matrix
Q = init_Q(n_actions, params)
Q_saved = Q.copy()

steps_per_state = init_steps_per_state()

# Visualisation
n_episodes_save = int(np.ceil(params['n_episodes'] / params['update_episode_division']))

# Initializing the reward
evo_training = {
    'evo_avg_reward_per_step': []
    , 'evo_n_steps': []
    , 'evo_avg_happiness': []
    , 'checking': get_default_checking()
}


# Training

# Starting the learning
for episode in tqdm(range(params['n_episodes'])):

    info_params_dict['n_episodes'] += 1
    info_params = get_info_params(info_params_dict)

    evo_episode = {
        'n_episode_steps': 0
        , 'done': False
        , 'episode_sum_reward': 0
        , 'evolution_sum_happiness': 0
    }

    # Get episode

    state1 = env_reset(params['start_at_random'])
    evo_episode['evolution_sum_happiness'] += get_happiness(state1)

    # Update parameters
    epsilon = get_epsilon(
        params_epsilon=params['epsilon'], episode=episode, steps_state=steps_per_state[state1['state_id']])
    evo_training['checking']['evo_epsilon'].append(epsilon)

    alpha = get_alpha(
        params_alpha=params['alpha'], episode=episode, steps_state=steps_per_state[state1['state_id']])
    evo_training['checking']['evo_alpha'].append(alpha)

    action1 = epsilon_greedy(Q, state1['state_id'], n_actions, epsilon)
    steps_per_state = update_steps_per_state(steps_per_state, state1['state_id'])

    while (not evo_episode['done']) and (evo_episode['n_episode_steps'] < params['nmax_steps']):

        # Getting the next state
        state2, reward1, evo_episode['done'], info = env_step(state1, action1)
        evo_episode['episode_sum_reward'] += reward1
        evo_episode['evolution_sum_happiness'] += get_happiness(state2)
        evo_training['checking']['empirical_return'].append(reward1)
        evo_training['checking']['predicted_value'].append(Q[state1['state_id']][action1])

        # Update parameters
        epsilon = get_epsilon(
            params_epsilon=params['epsilon'], episode=episode, steps_state=steps_per_state[state2['state_id']])
        evo_training['checking']['evo_epsilon'].append(epsilon)

        # Choosing the next action
        action2 = epsilon_greedy(Q, state2['state_id'], n_actions, epsilon)
        steps_per_state = update_steps_per_state(steps_per_state, state2['state_id'])

        # Learning the Q-value
        alpha = get_alpha(
            params_alpha=params['alpha'], episode=episode, steps_state=steps_per_state[state1['state_id']])
        method_params = {'alpha': alpha, 'gamma': params['gamma']}
        Q = update_Q_SARSA(Q, state1['state_id'], action1, reward1, state2['state_id'], action2, method_params)

        # Updating the respective values
        state1 = state2
        action1 = action2
        evo_episode['n_episode_steps'] += 1

    # At the end of learning process
    if params['render_episode']:
        print('Episode {}, Avg Reward : {}, Timesteps: {}, Epsilon: {}, Alpha: {}'.format(
            episode + 1,
            evo_episode['episode_sum_reward'] / evo_episode['n_episode_steps'],
            evo_episode['n_episode_steps'], epsilon, alpha))

    evo_training['evo_avg_reward_per_step'].append(evo_episode['episode_sum_reward'] / evo_episode['n_episode_steps'])
    evo_training['evo_n_steps'].append(evo_episode['n_episode_steps'])
    evo_training['evo_avg_happiness'].append(evo_episode['evolution_sum_happiness'] / evo_episode['n_episode_steps'])

    if (episode + 1) % n_episodes_save == 0:

        evo_training['checking'], Q_saved = launch_checking(evo_training['checking'], Q_saved, Q, method_id, info_params)
        save_models({'Q': Q, 'evo_training': evo_training}, method_id, final=False)

        plot_evolution_reward(evo_training['evo_avg_reward_per_step'], method_id, info_params)
        plot_evolution_steps(evo_training['evo_n_steps'], method_id, params['nmax_steps'], info_params)
        plot_evolution_happiness(evo_training['evo_avg_happiness'], method_id, info_params)
        # display_analyse_Q(Q, params['method'], info_params)


save_models({'Q': Q, 'evo_training': evo_training}, method_id, final=True)
plot_evolution_reward(evo_training['evo_avg_reward_per_step'], method_id, info_params)
plot_evolution_steps(evo_training['evo_n_steps'], method_id, params['nmax_steps'], info_params)
plot_evolution_happiness(evo_training['evo_avg_happiness'], method_id, info_params)
display_analyse_Q(Q, method_id, info_params)

import pandas as pd

analyse_Q = pd.DataFrame(columns=[
    'state_id', 'food_id', 'fat_id', 'affection_id', 'action_possible',
    'no_action', 'walking', 'feeding', 'playing',
])

for i in tqdm(Q.keys()):
    i_info = i.split('_')
    Q_info = Q[i]
    analyse_Q = analyse_Q.append({
        'state_id': i,
        'food_id': float(i_info[0]),
        'fat_id': float(i_info[1]),
        'affection_id': float(i_info[2]),
        'action_possible': i_info[3],
        'no_action': Q_info[0],
        'walking': Q_info[1],
        'feeding': Q_info[2],
        'playing': Q_info[3]
        }, ignore_index=True)


analyse_Q

import matplotlib.pyplot as plt
import statistics
from scipy import stats
import operator
import dill
import pandas as pd
from tqdm import tqdm
from src.utils.plots import COLORS
import numpy as np
from src.utils.maths import running_mean, get_avg_n_points


def display_state_characteristic(analyse_Q, state_characteristic, action_possible, method_id, info_params):

    dfgb = analyse_Q[analyse_Q['action_possible'] == action_possible].groupby(state_characteristic).agg({
        'no_action': 'mean',
        'walking': 'mean',
        'feeding': 'mean',
        'playing': 'mean'
    })
    dfgb.reset_index(inplace=True)

    fig = plt.figure()

    cnt = 0
    for action in ['no_action', 'walking', 'feeding', 'playing']:
        plt.plot(
            dfgb[state_characteristic], dfgb[action], label=action
            , marker='', color=COLORS[cnt], linewidth=1, alpha=0.75
        )
        cnt += 1

    plt.title('Avg Q-Value for state characteristic / action')
    plt.xlabel(state_characteristic + ' \n ' + info_params)
    plt.ylabel('Avg Q-Value')
    # plt.yscale('log')
    plt.xlim([-0.05, 1.05])
    plt.legend(bbox_to_anchor=(0.5, -0.10), loc="lower center",
               bbox_transform=fig.transFigure, ncol=4, fancybox=True, shadow=True, borderpad=1)
    plt.grid(True)
    plt.tight_layout()
    plt.savefig('data/figures/{}__Q_value_{}_{}.png'.format(method_id, state_characteristic, action_possible),
                format='png', dpi=1000, bbox_inches='tight')
    # plt.show()
    plt.close(fig)


for state_characteristic in ['food_id', 'fat_id', 'affection_id']:
    print(state_characteristic)
    display_state_characteristic(analyse_Q, state_characteristic, 'True', method_id, info_params)
    display_state_characteristic(analyse_Q, state_characteristic, 'False', method_id, info_params)

Q

def flatten(l):
    return [item for sublist in l for item in sublist]


def get_avg_Q(Q):

    q_Q = flatten(list(Q.values()))

    if len(q_Q) == 0:
        return 0
    else:
        return np.mean(q_Q)

get_avg_Q(Q)
