==> notebooks/6.0-yp-run_comparison.ipynb <==
import os
print(os.getcwd())
def update_working_directory():
    from pathlib import Path
    p = Path(os.getcwd()).parents[0]
    os.chdir(p)
    print(p)
update_working_directory()

import dill

from src.utils.plots import get_info_params

from src.visualization.rl_plots_comparison import plot_comparison_evolution_reward, plot_comparison_evolution_steps
from src.visualization.plot_happiness import plot_comparison_evolution_happiness

list_methods = ['sarsa_1', 'sarsa_2']

info_params = get_info_params({
    'method': list_methods
    , "n_episodes": 1000
    , 'nmax_steps': 1440
    , "gamma": 0.95
})

evo_training__evo_avg_reward_per_step = {}
evo_training__evo_n_steps = {}
evo_training__evo_avg_happiness = {}

for method in list_methods:
    print(method)
    with open("models/{}__evo_training.pkl".format(method), "rb") as input_file:
        evo_training = dill.load(input_file)

    evo_training__evo_avg_reward_per_step[method] = evo_training['evo_avg_reward_per_step']
    evo_training__evo_n_steps[method] = evo_training['evo_n_steps']
    evo_training__evo_avg_happiness[method] = evo_training['evo_avg_happiness']


print('Reward')
plot_comparison_evolution_reward(evo_training__evo_avg_reward_per_step, info_params)
print('Steps')
plot_comparison_evolution_steps(evo_training__evo_n_steps, nmax_steps, info_params)
print('Happiness')
plot_comparison_evolution_happiness(evo_training__evo_avg_happiness, info_params)

evolution_steps = [1,1,1,2,2,2,3,3,3,4,4,4,5,5,5]
evolution_steps

from src.utils.maths import get_avg_n_points
avg_n_points = get_avg_n_points(evolution_steps, n_points=5)
avg_n_points

n_points_batch = len(avg_n_points) - 2
n_points_batch

import numpy as np
n_points = 5
n_moving_points = int(np.ceil(len(evolution_steps) / n_points))
n_moving_points

for i_point_batch in range(n_points_batch):
    print(evolution_steps[i_point_batch*n_moving_points:(i_point_batch+1)*n_moving_points])

list_points = evolution_steps
n_points = 5

def get_min_n_points(list_points, n_points):
    n_moving_points = int(np.ceil(len(list_points) / n_points))
    min_n_points = []
    for i_point_batch in range(n_points):
        point_batch = evolution_steps[i_point_batch*n_moving_points:(i_point_batch+1)*n_moving_points]
        min_n_points.append(min(point_batch))
    return [list_points[0]] + min_n_points + [list_points[-1]]

def get_max_n_points(list_points, n_points):
    n_moving_points = int(np.ceil(len(list_points) / n_points))
    min_n_points = []
    for i_point_batch in range(n_points):
        point_batch = evolution_steps[i_point_batch*n_moving_points:(i_point_batch+1)*n_moving_points]
        min_n_points.append(max(point_batch))
    return [list_points[0]] + min_n_points + [list_points[-1]]

get_min_n_points(list_points, n_points)

get_max_n_points(list_points, n_points)
