==> notebooks/8.0-yp-run_evolution_episode.ipynb <==
import os
print(os.getcwd())
def update_working_directory():
    from pathlib import Path
    p = Path(os.getcwd()).parents[0]
    os.chdir(p)
    print(p)
update_working_directory()

import dill
import json
import argparse

from src.utils.plots import get_info_params

from src.visualization.rl_plots_comparison import plot_comparison_evolution_reward, plot_comparison_evolution_steps
from src.visualization.plot_happiness import plot_comparison_evolution_happiness, plot_comparison_evolution_episode


method_id_path = 'value_based/sarsa/sarsa_test'

method_id = method_id_path.split('/')[-1]
params = json.loads(open('src/models/run_evolution_episode.json').read())
params_method = json.loads(open('src/models/{}.json'.format(method_id_path)).read())

info_params_dict = {
    "nmax_steps": params['nmax_steps']
    , "gamma": params['gamma']
}
info_params = get_info_params(info_params_dict)

evo_episodes = []

for episode in range(-1,params['n_episodes'],10)[1:]:
    
    with open("data/interim/{}__evo_episode_{}.pkl".format(method_id, episode), "rb") as input_file:
        evo_episode = dill.load(input_file)

    evo_episodes.append({'episode': episode, 'evo_episode': evo_episode['happiness']})

plot_evolution_episode(evo_episodes, info_params)

dataDir = 'data/figures/evolution_episode/'
make_gif_evolution_episode(dataDir)
